// src/api/generated/schemas/getPostsFollow.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * ClothingFit API
 * ClothingFit 백엔드 API 문서
 * OpenAPI spec version: 1.0.0
 */

export type GetPostsFollow = typeof GetPostsFollow[keyof typeof GetPostsFollow];


export const GetPostsFollow = {
  TRUE: 'TRUE',
  FALSE: 'FALSE',
} as const;