// src/api/generated/schemas/getPostsParams.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * ClothingFit API
 * ClothingFit 백엔드 API 문서
 * OpenAPI spec version: 1.0.0
 */
import type { GetPostsFollow } from './getPostsFollow';
import type { GetPostsGender } from './getPostsGender';
import type { GetPostsSort } from './getPostsSort';

export type GetPostsParams = {
follow?: GetPostsFollow;
gender?: GetPostsGender;
sort?: GetPostsSort;
/**
 * @maxLength 100
 */
keyword?: string;
/**
 * @minimum 0
 * @maximum 300
 * @nullable
 */
height?: number | null;
/**
 * @minimum 0
 * @maximum 500
 * @nullable
 */
weightMin?: number | null;
/**
 * @minimum 0
 * @maximum 500
 * @nullable
 */
weightMax?: number | null;
cursor?: string;
/**
 * @minimum 1
 * @maximum 50
 */
limit?: number;
};